pyproject.toml
[project]
name = "berlin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.122.0",
    "google-genai>=1.52.0",
    "mss>=10.1.0",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "playwright>=1.56.0",
    "py-tgcalls>=2.2.8",
    "pyaudio>=0.2.14",
    "pyautogui>=0.9.54",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
    "telethon>=1.42.0",
]
